Actualizacion para que les funcione el front tienen que pull.. esto primero

+++ Backend/src/main/java/ElEcuipoGrupos30/Backend/DAO/ClientesDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ElEcuipoGrupos30.Backend.modelo.Clientes;
 
-public interface ClientesDAO extends JpaRepository<Clientes, Integer>{
+public interface ClientesDAO extends JpaRepository<Clientes, Long>{
 
 }
