+++ Backend/src/main/java/ElEcuipoGrupos30/Backend/DAO/ProductosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ElEcuipoGrupos30.Backend.modelo.Productos;
 
-public interface ProductosDAO extends JpaRepository<Productos, Integer>{
+public interface ProductosDAO extends JpaRepository<Productos, Long>{
 
 }
