+++ Backend/src/main/java/ElEcuipoGrupos30/Backend/DAO/ProveedoresDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ElEcuipoGrupos30.Backend.modelo.Proveedores;
 
-public interface ProveedoresDAO extends JpaRepository<Proveedores, Integer> {
+public interface ProveedoresDAO extends JpaRepository<Proveedores, Long> {
 
 }
