+++ Backend/src/main/java/ElEcuipoGrupos30/Backend/DAO/VentasDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ElEcuipoGrupos30.Backend.modelo.Ventas;
 
-public interface VentasDAO extends JpaRepository<Ventas, Integer> {
+public interface VentasDAO extends JpaRepository<Ventas, Long> {
 
 }
