+++ Backend/src/main/java/ElEcuipoGrupos30/Backend/api/ProductosAPI.java
@@ -29,7 +29,7 @@ public class ProductosAPI {
 	}
 	
 	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id") Integer id) {
+	public void eliminar(@PathVariable("id") Long id) {
 		productosDAO.deleteById(id);		
 	}
